some cleanup, marked one place in the code where we run into CS-347

+++ test/libs/uti/test_uti_string.cc
@@ -42,7 +42,7 @@ int main(int argc, char *argv[]) {
    bool ret = true;
 
    char buffer[10];
-   int len;
+   size_t len;
 
    sge_strlcpy(buffer, "12345678901234567890", sizeof(buffer));
    printf("%2d %s\n", (int) strlen(buffer), buffer);
@@ -53,21 +53,21 @@ int main(int argc, char *argv[]) {
 
    sge_strlcpy(buffer, "1234", sizeof(buffer));
    len = sge_strlcat(buffer, "1234", sizeof(buffer));
-   printf("%2d %s\n", len, buffer);
+   printf("%2zu %s\n", len, buffer);
    if (len != 9 || strcmp(buffer, "12341234") != 0) {
       fprintf(stderr, "sge_strlcat(1) failed\n");
       ret = false;
    }
 
    len = sge_strlcat(buffer, "1234", sizeof(buffer));
-   printf("%2d %s\n", len, buffer);
+   printf("%2zu %s\n", len, buffer);
    if (len != 13 || strcmp(buffer, "123412341") != 0) {
       fprintf(stderr, "sge_strlcat(1) failed\n");
       ret = false;
    }
 
    len = sge_strlcat(buffer, "1234", sizeof(buffer));
-   printf("%2d %s\n", len, buffer);
+   printf("%2zu %s\n", len, buffer);
    if (len != 14 || strcmp(buffer, "123412341") != 0) {
       fprintf(stderr, "sge_strlcat(1) failed\n");
       ret = false;
